@@ -45,8 +45,11 @@ def format_link_object_for_client(link, with_text, ref, pos=None):
     com["sourceHeRef"]       = linkRef.he_normal()
     com["anchorVerse"]       = anchorRef.sections[-1] if len(anchorRef.sections) else 0
     com["sourceHasEn"]       = "en" in linkLangs
-    com["anchorText"]        = getattr(link, "anchorText", "")
-    com["inline_reference"]  = getattr(link, "inline_reference", None)
+    # com["anchorText"]        = getattr(link, "anchorText", "") # not currently used
+    if getattr(link, "inline_reference", None):
+        com["inline_reference"]  = getattr(link, "inline_reference", None)
+    if getattr(link, "highlightedWords", None):
+        com["highlightedWords"] = getattr(link, "highlightedWords", None)
 
     compDate = getattr(linkRef.index, "compDate", None)
     if compDate:
