@@ -202,6 +202,9 @@ class UserHistory(abst.AbstractMongoRecord):
 class UserHistorySet(abst.AbstractMongoSet):
     recordClass = UserHistory
 
+    def hits(self):
+        return reduce(lambda agg,o: agg + getattr(o, "num_times_read", 1), self, 0)
+
 
 class UserProfile(object):
     def __init__(self, id=None, slug=None, email=None):
@@ -661,111 +664,6 @@ def public_user_data(uid):
     }
     public_user_data_cache[uid] = data
     return data
-
-# Properly, this doesn't belong here, but it's being developed alongside user_stats_data
-# todo: relocate the stats methods?
-def site_stats_data():
-    from sefaria.model.category import TOP_CATEGORIES
-    from sefaria.model.trend import Trend, TrendSet, read_in_category_key, reverse_read_in_category_key
-
-    d = {"categoriesRead": {reverse_read_in_category_key(t.name): t.value
-                            for t in TrendSet({"scope": "site", "period": "alltime",
-                                "name": {"$in": map(read_in_category_key, TOP_CATEGORIES)}
-                            })}}
-    return d
-
-
-def user_stats_data(uid, start=None, end=None):
-    """
-
-    :param uid: int or something cast-able to int
-    :param start: datetime
-    :param end: datetime
-    :return:
-    """
-    from sefaria.model.category import TOP_CATEGORIES
-    from sefaria.model.trend import Trend, TrendSet, read_in_category_key, reverse_read_in_category_key
-    from sefaria.sheets import user_sheets, sheet_list
-
-    uid = int(uid)
-
-    # todo: needs more thought.  UserHistory.timeclause handles the Nones, but later usages in this method aren't so graceful.
-    timeclause = UserHistory.timeclause(start, end)
-    end = end or datetime.now()
-    start = start or datetime(2017, 12, 1)  # start of Sefaria epoch
-
-    # All of user's sheets
-    usheets = user_sheets(uid)["sheets"]
-    usheet_ids = [s["id"] for s in usheets]
-
-    # Sheet views in this period
-    match_clause = {
-            "is_sheet": True,
-            "sheet_id": {"$in": usheet_ids},
-            }
-    match_clause.update(timeclause)
-    usheet_views = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$sheet_id",
-            "cnt": {"$sum": 1}}},
-    ])
-
-    most_popular_sheet_ids = [s["_id"] for s in sorted(usheet_views, key=lambda o: o["cnt"], reverse=True)[:3]]
-    most_popular_sheets = [s for s in usheets if s["id"] in most_popular_sheet_ids]
-    sheets_this_period = [s for s in usheets if start <= datetime.strptime(s["created"], "%Y-%m-%dT%H:%M:%S.%f") <= end]
-
-    # Refs I viewed
-    match_clause = {
-            "uid": uid,
-            "secondary": False,
-            "is_sheet": False
-            }
-    match_clause.update(timeclause)
-    refs_viewed = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$ref",
-            "cnt": {"$sum": 1}}},
-    ])
-    most_viewed_refs = [s["_id"] for s in sorted(refs_viewed, key=lambda o: o["cnt"], reverse=True) if s["cnt"] > 1 and "Genesis 1" not in s["_id"]][:10]
-
-
-    # Sheets I viewed
-    match_clause = {
-            "uid": uid,
-            "secondary": False,
-            "is_sheet": True
-            }
-    match_clause.update(timeclause)
-    sheets_viewed = db.user_history.aggregate([
-        {"$match": match_clause},
-        {"$group": {
-            "_id": "$sheet_id",
-            "cnt": {"$sum": 1}}},
-    ])
-    most_viewed_sheets_ids = [s["_id"] for s in sorted(sheets_viewed, key=lambda o: o["cnt"], reverse=True) if s["cnt"] > 1 and s["_id"] not in usheet_ids][:10]
-    most_viewed_sheets = sheet_list({"id": {"$in":most_viewed_sheets_ids}})
-
-    # Construct returned data
-    d = public_user_data(uid)
-
-    sheetsReadQuery = {"is_sheet": True, "secondary": False, "uid": uid}
-    sheetsReadQuery.update(timeclause)
-    d["sheetsRead"] = UserHistorySet(sheetsReadQuery).count()
-
-    textsReadQuery = {"is_sheet": False, "secondary": False, "uid": uid}
-    textsReadQuery.update(timeclause)
-    d["textsRead"] = UserHistorySet(textsReadQuery).count()
-
-    d["categoriesRead"] = {reverse_read_in_category_key(t.name): t.value for t in TrendSet({"uid":uid, "name": {"$in": map(read_in_category_key, TOP_CATEGORIES)}})}
-    d["totalSheets"] = len(usheets)
-    d["publicSheets"] = len([s for s in usheets if s["status"] == "public"])
-    d["popularSheets"] = most_popular_sheets
-    d["sheetsThisPeriod"] = len(sheets_this_period)
-    d["mostViewedRefs"] = most_viewed_refs
-    d["mostViewedSheets"] = most_viewed_sheets
-    return d
 
 
 def user_name(uid):
