@@ -390,7 +390,7 @@
                 ns._wrapMatches();
 
                 if (ns.matches.length == 0) {
-                    console.log("No references found to link to Sefaria.");
+                    //console.log("No references found to link to Sefaria.");
                     return;
                 }
 
@@ -475,6 +475,9 @@
     }
 
     ns._trackPage = function() {
+        var robots = document.head.querySelector("meta[name~=robots]");
+        if (robots && robots.content.includes("noindex")) { return; }
+
         var canonical = document.head.querySelector("link[rel~=canonical]");
         var url = canonical ? canonical.href : window.location.href;
         var meta = document.head.querySelector("meta[name~=description]")
@@ -490,6 +493,7 @@
             "description": description,
             "refs": ns.matches,
         };
+        //console.log("TRACK");
         //console.log(data);
         var json = JSON.stringify(data);
         var postData = encodeURIComponent("json") + '=' + encodeURIComponent(json);
